Lets use int for now, as we do have old databases created with INTs instead of double. Will be changed in near future.

+++ src/SoundFingerprinting.MongoDb/Entity/Track.cs
@@ -16,7 +16,7 @@
 
         public int ReleaseYear { get; set; }
 
-        public double TrackLengthSec { get; set; }
+        public int TrackLengthSec { get; set; }
 
         public string GroupId { get; set; }
     }
